clean up node modules utils and refactor polyfills plugin to not hardcode node_modules location
