+++ packages/plugin-polyfills/test/cases/default/default.spec.js
@@ -13,9 +13,8 @@
  *
  * {
  *   plugins: [{
- *     ...greenwoodPluginPolyfills()
- *  }]
- *
+ *     greenwoodPluginPolyfills()
+ *   }]
  * }
  *
  * User Workspace
