+++ packages/plugin-polyfills/test/cases/default/greenwood.config.js
@@ -2,6 +2,6 @@ import { greenwoodPluginPolyfills } from '../../../src/index.js';
 
 export default {
   plugins: [
-    ...greenwoodPluginPolyfills()
+    greenwoodPluginPolyfills()
   ]
 };
