+++ packages/plugin-polyfills/test/cases/dsd/dsd.spec.js
@@ -11,7 +11,7 @@
  * User Config
  * {
  *   plugins: [
- *     ...greenwoodPluginPolyfills({
+ *     greenwoodPluginPolyfills({
  *       dsd: true
  *     })
  *   ]
