+++ packages/plugin-polyfills/test/cases/dsd/greenwood.config.js
@@ -2,7 +2,7 @@ import { greenwoodPluginPolyfills } from '../../../src/index.js';
 
 export default {
   plugins: [
-    ...greenwoodPluginPolyfills({
+    greenwoodPluginPolyfills({
       dsd: true,
       wc: false
     })
