+++ packages/plugin-polyfills/test/cases/lit/lit.spec.js
@@ -13,7 +13,7 @@
  *
  * {
  *   plugins: [
- *     ...greenwoodPluginPolyfills({
+ *     greenwoodPluginPolyfills({
  *       lit: true
  *     })
  *   ]
